requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
PyJWT==2.8.0
werkzeug==2.3.0
gunicorn==21.0.0
